requirements.txt
Django==1.4.5
South==0.7.6
distribute==0.6.27
dj-database-url==0.2.1
djangorestframework==2.2.1
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
six==1.2.0
wsgiref==0.1.2
